Remove `ExpandTo` class
