+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/LocationCreator.scala
@@ -2,9 +2,10 @@ package io.shiftleft.semanticcpg.language
 
 import io.shiftleft.codepropertygraph.generated.nodes
 import io.shiftleft.codepropertygraph.generated.nodes.NewLocation
-import io.shiftleft.semanticcpg.utils.ExpandTo
 import org.slf4j.{Logger, LoggerFactory}
 import overflowdb.traversal._
+
+import scala.annotation.tailrec
 
 /* TODO MP: this should be part of the normal steps, rather than matching on the type at runtime
  * all (and only) steps extending DataFlowObject should/must have `newSink`, `newSource` and `newLocation` */
@@ -122,7 +123,7 @@ object LocationCreator {
     if (method == null) {
       nodes.NewLocation("", "", "", "", None, "", "", "", "", Some(node))
     } else {
-      val typeOption = ExpandTo.methodToTypeDecl(method)
+      val typeOption = methodToTypeDecl(method)
       val typeName = typeOption.map(_.fullName).getOrElse("")
       val typeShortName = typeOption.map(_.name).getOrElse("")
 
@@ -147,6 +148,17 @@ object LocationCreator {
       )
     }
   }
+
+  private def methodToTypeDecl(method: nodes.Method): Option[nodes.TypeDecl] =
+    findVertex(method, _.isInstanceOf[nodes.TypeDecl]).map(_.asInstanceOf[nodes.TypeDecl])
+
+  @tailrec
+  private def findVertex(node: nodes.StoredNode, instanceCheck: nodes.StoredNode => Boolean): Option[nodes.StoredNode] =
+    node._astIn.nextOption match {
+      case Some(head) if instanceCheck(head) => Some(head)
+      case Some(head)                        => findVertex(head, instanceCheck)
+      case None                              => None
+    }
 
   def emptyLocation(label: String, node: Option[nodes.CpgNode]): nodes.NewLocation =
     nodes.NewLocation("", "", "", "", None, "", "", label, "", node)
